windows: the core, plugins, and utils now compile with mingw.
That doesn't mean they *work*, but they do compile.  The filesystem
stuff in codec.c still doesn't work, this basically just #ifdefs all
the problematic stuff out.  Benchmark also doesn't work yet because I
need to create a Windows alternative for the timer.

Partially fixes https://github.com/quixdb/squash/issues/38

+++ examples/simple.c
@@ -37,8 +37,8 @@ int main (int argc, char** argv) {
     exit (res);
   }
 
-  fprintf (stdout, "Compressed a %zu byte buffer to %zu bytes.\n",
-           uncompressed_length, compressed_length);
+  fprintf (stdout, "Compressed a %u byte buffer to %u bytes.\n",
+           (unsigned int) uncompressed_length, (unsigned int) compressed_length);
 
   res = squash_decompress (codec_name,
                            (uint8_t*) decompressed, &decompressed_length,
